@@ -13,7 +13,8 @@
   ],
   "minimum-stability": "stable",
   "require": {
-    "php": "^8.1.4"
+    "php": "^8.2",
+    "sinergi/browser-detector": "^6.1"
   },
   "require-dev": {
     "phpunit/phpunit": "^9.5"
